Added a secure-by-default HTTP client for databases.
This client is used for NMDC but not JDP, since the latter doesn't (yet) support
HTTP Strict Transport Security (HSTS). Probably a good idea to encourage them to
support this feature in the near future.
